% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_stan_package.R
\name{stan.ridge.plots.f0}
\alias{stan.ridge.plots.f0}
\title{Plot ridges for a latent variable with no index}
\usage{
stan.ridge.plots.f0(latent_var)
}
\arguments{
\item{latent_var}{The number of colours to be produced}
}
\value{
density plot (using ggplot2)
}
\description{
This function plots a density plot for a latent variable in stan that has no index.
For example a variable in stan written "tau ~ normal(0,10)"
}
\examples{
stan.ridge.plots.f0("tau")

}
