% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_stan_package.R
\name{stan_param_collapse_1f}
\alias{stan_param_collapse_1f}
\title{Collapse the latent variable to 1 index}
\usage{
stan_param_collapse_1f(
  latent_var,
  f1_colname,
  idx_f1_colname,
  input_data,
  i = F
)
}
\arguments{
\item{latent_var}{The name of the variable (i.e. "mu")}

\item{f1_colname}{The column name where the factor is as a vector (i.e. "sample_names")}

\item{idx_f1_colname}{The column name where the index is as a vector (i.e. "sample_idx")}

\item{input_data}{The dataframe where all the colums above exists}

\item{i}{Logical vector for stating the indexing of the parameters needs adjusting}
}
\value{
arranged latent parameter output from stan
}
\description{
This function merges the latent variable from stan output onto its given index and prepares
the data for ridged plots. Example of a latent variable written in stan is "mu\link{sample_idx}".
}
\examples{
stan_param_collapse_1f("sigma","species_name","species_idx",qpcr_df,i=F)
stan.plot.data <- stan_param_collapse_1f("phi_1","species_name","species_idx",st_qpcr)
}
